feat(toolkit): render sheared glyph runs for italic text
The TTF rasterizer had no italic face, so italic spans drew upright and were
indistinguishable from normal text. draw_text_sheared applies a horizontal
shear proportional to the distance above the baseline, giving a synthetic
oblique at the OBLIQUE constant, and returns the advance so callers keep
measuring runs the same way.
