+++ userland/toolkit/src/font/ttf/mod.rs
@@ -25,9 +25,11 @@ mod draw;
 mod face;
 mod metrics;
 mod readable;
+mod slant;
 
 pub use face::builtin_face;
 pub use readable::MIN_UI_PX;
+pub use slant::{draw_text_sheared, OBLIQUE};
 
 pub use ab_glyph::FontRef;
 pub use draw::{draw_text, draw_text_spaced, draw_text_tracked, draw_text_with};
